feature (ui): add sass and rollup support

+++ src/utils/dateHelper.js
@@ -1,6 +1,6 @@
 export const timeFormatter = isoTime => {
     let date = new Date(isoTime)
-    let options = { year: 'numeric', month: 'long', day: '2-digit'}
+    let options = { year: 'numeric', month: 'short', day: '2-digit'}
     return date.toLocaleDateString('es-ES', options)
 }
